added new dependencies (glm, glew). The CalcPath method in DriveCOmplex has been re-written

+++ src/game/struct/VehiclePropetries.hpp
@@ -26,7 +26,7 @@ struct VehiclePropetries
 
 	int protection;
 	int radar;
-	int speed;  // depends on mass and drive
+	float speed;  // depends on mass and drive
 
 	int energy;
 	int hyper;  // depends on drive and bak
